Switch to using Module.prepend instead of method aliasing in redis hook
